+++ ORStools/proc/matrix_proc.py
@@ -88,7 +88,8 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
         # Get profile value
         profile = dict(enumerate(PROFILES))[parameters[self.IN_PROFILE]]
 
-        options = self.parseOptions(parameters, context)
+        # TODO: enable once core matrix is available
+        # options = self.parseOptions(parameters, context)
 
         # Get parameter values
         source = self.parameterAsSource(
@@ -146,8 +147,8 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             'sources': sources_ids,
             'destinations': destination_ids,
             'metrics': ["duration", "distance"],
-            'id': 'Matrix',
-            'options': options
+            'id': 'Matrix'
+            # 'options': options
         }
 
         # get types of set ID fields